+++ mrpt_reactivenav2d/CMakeLists.txt
@@ -1,27 +1,34 @@
-cmake_minimum_required(VERSION 3.1)
+cmake_minimum_required(VERSION 3.5)
 project(mrpt_reactivenav2d)
 
-## Find catkin macros and libraries
-## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
-## is used, also find other catkin packages
-find_package(catkin REQUIRED COMPONENTS
-  actionlib
-  actionlib_msgs
-  dynamic_reconfigure
-  geometry_msgs
-  roscpp
-  tf2
-  tf2_ros
-  tf2_geometry_msgs
-  visualization_msgs
-)
+# find dependencies
+find_package(ament_cmake REQUIRED)
+find_package(rclcpp REQUIRED)
+find_package(rclcpp_components REQUIRED)
+find_package(nav_msgs REQUIRED)
+find_package(geometry_msgs)
+find_package(sensor_msgs REQUIRED)
+find_package(tf2 REQUIRED)
+find_package(tf2_ros)
+find_package(tf2_geometry_msgs REQUIRED)
+find_package(visualization_msgs REQUIRED)
+find_package(actionlib)
+find_package(actionlib_msgs)
+
 
 ## System dependencies are found with CMake's conventions
-find_package(mrpt-ros1bridge REQUIRED)
+find_package(mrpt-ros2bridge REQUIRED)
 find_package(mrpt-nav REQUIRED)
 find_package(mrpt-kinematics REQUIRED)
 
 message(STATUS "MRPT_VERSION: ${MRPT_VERSION}")
+if(NOT CMAKE_C_STANDARD)
+  set(CMAKE_C_STANDARD 99)
+endif()
+
+if(NOT CMAKE_CXX_STANDARD)
+  set(CMAKE_CXX_STANDARD 17)
+endif()
 
 if (CMAKE_COMPILER_IS_GNUCXX)
 	# High level of warnings.
@@ -37,60 +44,42 @@ IF(CMAKE_COMPILER_IS_GNUCXX AND NOT CMAKE_BUILD_TYPE MATCHES "Debug")
   add_compile_options(-O3)
 ENDIF()
 
-###################################
-## catkin specific configuration ##
-###################################
-## The catkin_package macro generates cmake config files for your package
-## Declare things to be passed to dependent projects
-## INCLUDE_DIRS: uncomment this if you package contains header files
-## LIBRARIES: libraries you create in this project that dependent projects also need
-## CATKIN_DEPENDS: catkin_packages dependent projects also need
-## DEPENDS: system dependencies of this project that dependent projects also need
-catkin_package(
-#  INCLUDE_DIRS include
-#  LIBRARIES mrpt_reactivenav2d
-  CATKIN_DEPENDS
-    actionlib
-    actionlib_msgs
-    dynamic_reconfigure
-    geometry_msgs
-    roscpp
-    tf2
-    tf2_ros
-    tf2_geometry_msgs
-      visualization_msgs
-#  DEPENDS mrpt
-)
-
 ###########
 ## Build ##
 ###########
 
-## Specify additional locations of header files
-## Your package locations should be listed before other locations
-# include_directories(include)
-include_directories(
-  ${catkin_INCLUDE_DIRS}
+## Declare a cpp executable
+add_executable(${PROJECT_NAME}_node 
+                src/mrpt_reactivenav2d_node.cpp
+                include/${PROJECT_NAME}/mrpt_reactivenav2d_node.hpp
 )
 
-## Declare a cpp library
-# add_library(mrpt_reactivenav2d
-#   src/${PROJECT_NAME}/mrpt_reactivenav2d.cpp
-# )
-
-## Declare a cpp executable
-add_executable(mrpt_reactivenav2d_node src/mrpt_reactivenav2d_node.cpp)
-
-## Add cmake target dependencies of the executable/library
-## as an example, message headers may need to be generated before nodes
-# add_dependencies(mrpt_reactivenav2d_node mrpt_reactivenav2d_generate_messages_cpp)
+target_include_directories(${PROJECT_NAME}_node
+                          PUBLIC
+                          $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+                          $<INSTALL_INTERFACE:include>
+)               
 
 ## Specify libraries to link a library or executable target against
-target_link_libraries(mrpt_reactivenav2d_node
-  ${catkin_LIBRARIES}
+target_link_libraries(${PROJECT_NAME}_node
   mrpt::nav
   mrpt::kinematics
-  mrpt::ros1bridge
+  mrpt::ros2bridge
+)
+
+ament_target_dependencies(
+  ${PROJECT_NAME}_node
+  "rclcpp"
+  "rclcpp_components"
+  "nav_msgs"
+  "sensor_msgs"
+  "geometry_msgs"
+  "actionlib"
+  "actionlib_msgs"
+  "tf2"
+  "tf2_ros"
+  "tf2_geometry_msgs"
+  "visualization_msgs"
 )
 
 #############
@@ -102,19 +91,31 @@ target_link_libraries(mrpt_reactivenav2d_node
 
 
 ## Mark executables and/or libraries for installation
-install(TARGETS mrpt_reactivenav2d_node
-  ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+install(TARGETS ${PROJECT_NAME}_node
+  ARCHIVE DESTINATION lib/${PROJECT_NAME}
+  LIBRARY DESTINATION lib/${PROJECT_NAME}
+  RUNTIME DESTINATION bin/${PROJECT_NAME}
 )
 
 ## Mark other files for installation (e.g. launch and bag files, etc.)
 install(DIRECTORY
   launch
   tutorial
-  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
+  DESTINATION share/${PROJECT_NAME}
 )
 
 #############
 ## Testing ##
 #############
+if(BUILD_TESTING)
+  find_package(ament_lint_auto REQUIRED)
+  # the following line skips the linter which checks for copyrights
+  # uncomment the line when a copyright and license is not present in all source files
+  #set(ament_cmake_copyright_FOUND TRUE)
+  # the following line skips cpplint (only works in a git repo)
+  # uncomment the line when this package is not in a git repo
+  #set(ament_cmake_cpplint_FOUND TRUE)
+  ament_lint_auto_find_test_dependencies()
+endif()
+
+ament_package()
