+++ mrpt_reactivenav2d/include/mrpt_reactivenav2d/mrpt_reactivenav2d_node.hpp
@@ -1,0 +1,256 @@
+/***********************************************************************************
+ * Revised BSD License *
+ * Copyright (c) 2014-2023, Jose-Luis Blanco <jlblanco@ual.es> *
+ * All rights reserved. *
+ *                                                                                 *
+ * Redistribution and use in source and binary forms, with or without *
+ * modification, are permitted provided that the following conditions are met: *
+ *     * Redistributions of source code must retain the above copyright *
+ *       notice, this list of conditions and the following disclaimer. *
+ *     * Redistributions in binary form must reproduce the above copyright *
+ *       notice, this list of conditions and the following disclaimer in the *
+ *       documentation and/or other materials provided with the distribution. *
+ *     * Neither the name of the Vienna University of Technology nor the *
+ *       names of its contributors may be used to endorse or promote products *
+ *       derived from this software without specific prior written permission. *
+ *                                                                                 *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ *AND *
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ **
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE *
+ * DISCLAIMED. IN NO EVENT SHALL Markus Bader BE LIABLE FOR ANY *
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES *
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ **
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND *
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT *
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ **
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. *
+ ***********************************************************************************/
+
+
+#include <mrpt/config/CConfigFile.h>
+#include <mrpt/config/CConfigFileMemory.h>
+#include <mrpt/kinematics/CVehicleVelCmd_DiffDriven.h>
+#include <mrpt/maps/CSimplePointsMap.h>
+#include <mrpt/nav/reactive/CReactiveNavigationSystem.h>
+#include <mrpt/ros2bridge/point_cloud.h>
+#include <mrpt/ros2bridge/pose.h>
+#include <mrpt/ros2bridge/time.h>
+#include <mrpt/system/CTimeLogger.h>
+#include <mrpt/system/filesystem.h>
+
+#include <rclcpp/rclcpp.hpp>
+#include <sensor_msgs/msg/laser_scan.hpp>
+#include <sensor_msgs/msg/point_cloud.hpp>
+#include <geometry_msgs/msg/polygon.hpp>
+#include <geometry_msgs/msg/pose_stamped.hpp>
+#include <nav_msgs/msg/odometry.hpp>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#include <tf2_ros/buffer.h>
+#include <tf2_ros/transform_listener.h>
+
+#include <mutex>
+
+using namespace mrpt::nav;
+using mrpt::maps::CSimplePointsMap;
+using namespace mrpt::system;
+using namespace mrpt::config;
+
+// The ROS node
+class ReactiveNav2DNode: public rclcpp::Node
+{
+    public: 
+    /* Ctor*/
+    explicit ReactiveNav2DNode(const rclcpp::NodeOptions& options = rclcpp::NodeOptions());
+    /* Dtor*/
+    ~ReactiveNav2DNode(){}
+
+   private:
+   // methods 
+   void read_parameters();
+   void navigate_to(const mrpt::math::TPose2D& target);
+   void on_do_navigation();
+   void on_goal_received(const geometry_msgs::msg::PoseStamped::SharedPtr& trg_ptr);
+   void on_local_obstacles(const sensor_msgs::msg::PointCloud::SharedPtr& obs);
+   void on_set_robot_shape(const geometry_msgs::msg::Polygon::SharedPtr& newShape);
+
+	CTimeLogger m_profiler;
+	TAuxInitializer m_auxinit;	//!< Just to make sure ROS is init first
+	ros::NodeHandle m_nh;  //!< The node handle
+	ros::NodeHandle m_localn;  //!< "~"
+
+	/** @name ROS pubs/subs
+	 *  @{ */
+	ros::Subscriber m_sub_nav_goal;
+	ros::Subscriber m_sub_local_obs;
+	ros::Subscriber m_sub_robot_shape;
+	ros::Publisher m_pub_cmd_vel;
+
+	tf2_ros::Buffer m_tf_buffer;
+	tf2_ros::TransformListener m_tf_listener{m_tf_buffer};
+	/** @} */
+
+	bool m_1st_time_init;  //!< Reactive initialization done?
+	double m_target_allowed_distance;
+	double m_nav_period;
+
+	std::string m_pub_topic_reactive_nav_goal;
+	std::string m_sub_topic_local_obstacles;
+	std::string m_sub_topic_robot_shape;
+
+	std::string m_frameid_reference;
+	std::string m_frameid_robot;
+
+	bool m_save_nav_log;
+
+	ros::Timer m_timer_run_nav;
+
+	CSimplePointsMap m_last_obstacles;
+	std::mutex m_last_obstacles_cs;
+
+	struct MyReactiveInterface : public mrpt::nav::CRobot2NavInterface
+	{
+		ReactiveNav2DNode& m_parent;
+
+		MyReactiveInterface(ReactiveNav2DNode& parent) : m_parent(parent) {}
+
+		/** Get the current pose and speeds of the robot.
+		 *   \param curPose Current robot pose.
+		 *   \param curV Current linear speed, in meters per second.
+		 *	 \param curW Current angular speed, in radians per second.
+		 * \return false on any error.
+		 */
+		bool getCurrentPoseAndSpeeds(
+			mrpt::math::TPose2D& curPose, mrpt::math::TTwist2D& curVel,
+			mrpt::system::TTimeStamp& timestamp,
+			mrpt::math::TPose2D& curOdometry, std::string& frame_id) override
+		{
+			double curV, curW;
+
+			CTimeLoggerEntry tle(
+				m_parent.m_profiler, "getCurrentPoseAndSpeeds");
+
+			ros::Duration timeout(0.1);
+
+			geometry_msgs::TransformStamped tfGeom;
+			try
+			{
+				CTimeLoggerEntry tle2(
+					m_parent.m_profiler,
+					"getCurrentPoseAndSpeeds.lookupTransform_sensor");
+
+				tfGeom = m_parent.m_tf_buffer.lookupTransform(
+					m_parent.m_frameid_reference, m_parent.m_frameid_robot,
+					ros::Time(0), timeout);
+			}
+			catch (const tf2::TransformException& ex)
+			{
+				ROS_ERROR("%s", ex.what());
+				return false;
+			}
+
+			tf2::Transform txRobotPose;
+			tf2::fromMsg(tfGeom.transform, txRobotPose);
+
+			const mrpt::poses::CPose3D curRobotPose =
+				mrpt::ros1bridge::fromROS(txRobotPose);
+
+			timestamp = mrpt::ros1bridge::fromROS(tfGeom.header.stamp);
+
+			// Explicit 3d->2d to confirm we know we're losing information
+			curPose = mrpt::poses::CPose2D(curRobotPose).asTPose();
+			curOdometry = curPose;
+
+			curV = curW = 0;
+			MRPT_TODO("Retrieve current speeds from /odom topic?");
+			ROS_DEBUG(
+				"[getCurrentPoseAndSpeeds] Latest pose: %s",
+				curPose.asString().c_str());
+
+			// From local to global:
+			curVel = mrpt::math::TTwist2D(curV, .0, curW).rotated(curPose.phi);
+
+			return true;
+		}
+
+		/** Change the instantaneous speeds of robot.
+		 *   \param v Linear speed, in meters per second.
+		 *	 \param w Angular speed, in radians per second.
+		 * \return false on any error.
+		 */
+		bool changeSpeeds(
+			const mrpt::kinematics::CVehicleVelCmd& vel_cmd) override
+		{
+			using namespace mrpt::kinematics;
+			const CVehicleVelCmd_DiffDriven* vel_cmd_diff_driven =
+				dynamic_cast<const CVehicleVelCmd_DiffDriven*>(&vel_cmd);
+			ASSERT_(vel_cmd_diff_driven);
+
+			const double v = vel_cmd_diff_driven->lin_vel;
+			const double w = vel_cmd_diff_driven->ang_vel;
+			ROS_DEBUG(
+				"changeSpeeds: v=%7.4f m/s  w=%8.3f deg/s", v,
+				w * 180.0f / M_PI);
+			geometry_msgs::Twist cmd;
+			cmd.linear.x = v;
+			cmd.angular.z = w;
+			m_parent.m_pub_cmd_vel.publish(cmd);
+			return true;
+		}
+
+		bool stop(bool isEmergency) override
+		{
+			mrpt::kinematics::CVehicleVelCmd_DiffDriven vel_cmd;
+			vel_cmd.lin_vel = 0;
+			vel_cmd.ang_vel = 0;
+			return changeSpeeds(vel_cmd);
+		}
+
+		/** Start the watchdog timer of the robot platform, if any.
+		 * \param T_ms Period, in ms.
+		 * \return false on any error. */
+		virtual bool startWatchdog(float T_ms) override { return true; }
+		/** Stop the watchdog timer.
+		 * \return false on any error. */
+		virtual bool stopWatchdog() override { return true; }
+		/** Return the current set of obstacle points.
+		 * \return false on any error. */
+		bool senseObstacles(
+			CSimplePointsMap& obstacles,
+			mrpt::system::TTimeStamp& timestamp) override
+		{
+			timestamp = mrpt::system::now();
+			std::lock_guard<std::mutex> csl(m_parent.m_last_obstacles_cs);
+			obstacles = m_parent.m_last_obstacles;
+
+			MRPT_TODO("TODO: Check age of obstacles!");
+			return true;
+		}
+
+		mrpt::kinematics::CVehicleVelCmd::Ptr getEmergencyStopCmd() override
+		{
+			return getStopCmd();
+		}
+		mrpt::kinematics::CVehicleVelCmd::Ptr getStopCmd() override
+		{
+			mrpt::kinematics::CVehicleVelCmd::Ptr ret =
+				mrpt::kinematics::CVehicleVelCmd::Ptr(
+					new mrpt::kinematics::CVehicleVelCmd_DiffDriven);
+			ret->setToStop();
+			return ret;
+		}
+
+		void sendNavigationStartEvent() override {}
+		void sendNavigationEndEvent() override {}
+		void sendNavigationEndDueToErrorEvent() override {}
+		void sendWaySeemsBlockedEvent() override {}
+	};
+
+	MyReactiveInterface m_reactive_if;
+
+	CReactiveNavigationSystem m_reactive_nav_engine;
+	std::mutex m_reactive_nav_engine_cs;
+};
